clean up, make filter logic reusable
